requirements.txt
pyserial
pydub